move firebolt client to client package
