+++ acceptance_test.go
@@ -24,9 +24,9 @@ import (
 	"time"
 
 	"github.com/brianvoe/gofakeit"
+	"github.com/conduitio-labs/conduit-connector-firebolt/client"
 	"github.com/conduitio-labs/conduit-connector-firebolt/config"
 	"github.com/conduitio-labs/conduit-connector-firebolt/destination"
-	"github.com/conduitio-labs/conduit-connector-firebolt/firebolt"
 	"github.com/conduitio-labs/conduit-connector-firebolt/source"
 	sdk "github.com/conduitio/conduit-connector-sdk"
 	"github.com/matryer/is"
@@ -130,9 +130,9 @@ func prepareConfig(t *testing.T) map[string]string {
 }
 
 func prepareData(ctx context.Context, t *testing.T, cfg map[string]string) error {
-	cl := firebolt.NewClient(ctx, cfg[config.KeyDB])
+	cl := client.NewClient(ctx, cfg[config.KeyDB])
 
-	err := cl.Login(ctx, firebolt.LoginParams{
+	err := cl.Login(ctx, client.LoginParams{
 		Email:       cfg[config.KeyEmail],
 		Password:    cfg[config.KeyPassword],
 		AccountName: cfg[config.KeyAccountName],
