+++ source/source_integration_test.go
@@ -22,8 +22,8 @@ import (
 	"reflect"
 	"testing"
 
+	"github.com/conduitio-labs/conduit-connector-firebolt/client"
 	"github.com/conduitio-labs/conduit-connector-firebolt/config"
-	"github.com/conduitio-labs/conduit-connector-firebolt/firebolt"
 	sdk "github.com/conduitio/conduit-connector-sdk"
 )
 
@@ -173,9 +173,9 @@ func prepareConfig() (map[string]string, error) {
 }
 
 func prepareData(ctx context.Context, cfg map[string]string) error {
-	cl := firebolt.NewClient(ctx, cfg[config.KeyDB])
+	cl := client.NewClient(ctx, cfg[config.KeyDB])
 
-	err := cl.Login(ctx, firebolt.LoginParams{
+	err := cl.Login(ctx, client.LoginParams{
 		Email:       cfg[config.KeyEmail],
 		Password:    cfg[config.KeyPassword],
 		AccountName: cfg[config.KeyAccountName],
@@ -201,9 +201,9 @@ func prepareData(ctx context.Context, cfg map[string]string) error {
 }
 
 func clearData(ctx context.Context, cfg map[string]string) error {
-	cl := firebolt.NewClient(ctx, cfg[config.KeyDB])
+	cl := client.NewClient(ctx, cfg[config.KeyDB])
 
-	err := cl.Login(ctx, firebolt.LoginParams{
+	err := cl.Login(ctx, client.LoginParams{
 		Email:       cfg[config.KeyEmail],
 		Password:    cfg[config.KeyPassword],
 		AccountName: cfg[config.KeyAccountName],
@@ -223,9 +223,9 @@ func clearData(ctx context.Context, cfg map[string]string) error {
 }
 
 func prepareEmptyTable(ctx context.Context, cfg map[string]string) error {
-	cl := firebolt.NewClient(ctx, cfg[config.KeyDB])
+	cl := client.NewClient(ctx, cfg[config.KeyDB])
 
-	err := cl.Login(ctx, firebolt.LoginParams{
+	err := cl.Login(ctx, client.LoginParams{
 		Email:       cfg[config.KeyEmail],
 		Password:    cfg[config.KeyPassword],
 		AccountName: cfg[config.KeyAccountName],
